pages(dashboard): swap dashboard and dashboard_thin
