+++ src/pages/dashboard_thin.jsx
@@ -1,12 +1,35 @@
 import Layout from '@theme/Layout';
-import Dashboard from '@site/src/components/Dashboard';
+import ServiceData from '@site/src/components/ServiceData';
+import { translate } from "@docusaurus/Translate";
 
-const Dashboard_thin = () => {
+const PageHeader = ({ title }) => (
+  <header className="text-center pt-14 px-8 pb-6 border-b border-black/5">
+    <h1 className="text-3xl font-extrabold text-blue-900">
+      {title}
+    </h1>
+  </header>
+);
+
+const Dashboard = () => {
   return (
     <Layout title="Data Panel" description="RuyiSDK Data Panel">
-      <Dashboard />
+      <main className="min-h-screen flex flex-col font-sans bg-gray-100 text-gray-900">
+        <PageHeader
+          title={translate({
+            id: "dashboard.title",
+            message: "RuyiSDK 数据统计"
+          })}
+        />
+
+        <section className="flex-1 flex justify-center pt-10 pb-16 px-6">
+          <div className="w-full max-w-screen-xl mx-auto">
+            <ServiceData />
+          </div>
+        </section>
+
+      </main>
     </Layout>
   );
 };
 
-export default Dashboard_thin;
+export default Dashboard;
